🤗🚩 Updated with Glitch - Lessons 9-13.
