+++ package.json
@@ -14,12 +14,13 @@
     "express": "^4.16.4",
     "cors": "^2.8.5",
     "body-parser": "^1.18.3",
-    "mongodb": "^3.1.10",
+    "mongodb": "^2.2.33",
     "helmet": "^3.15.0",
     "pug": "^2.0.3",
     "passport": "^0.4.0",
     "express-session": "^1.15.6",
-    "passport-local": "^1.0.0"
+    "passport-local": "^1.0.0",
+    "bcrypt": "^3.0.3"
   },
   "engines": {
     "node": "4.4.3"
