+++ views/pug/profile.pug
@@ -11,6 +11,11 @@ html
     h1.border.center FCC Advanced Node and Express
     //add your code below, make sure its indented at this level
     
+    //#9 How to Put a Profile Together:
+    h2.center#welcome Welcome, #{username}!
+    
+    //#9 How to Put a Profile Together:
+    a(href='/logout') Logout
         
     script(src='https://code.jquery.com/jquery-2.2.1.min.js', integrity='sha256-gvQgAFzTH6trSrAWoH1iPo9Xc96QxSZ3feW6kem+O00=', crossorigin='anonymous')
     script(src='/public/client.js')
